Refactor collect into separate file
