+++ packages/transformers/js/core/src/fs.rs
@@ -1,5 +1,5 @@
 use crate::dependency_collector::{DependencyDescriptor, DependencyKind};
-use crate::hoist::{Collect, Import};
+use crate::hoist_collect::{HoistCollect, Import};
 use crate::utils::SourceLocation;
 use data_encoding::{BASE64, HEXLOWER};
 use std::collections::HashSet;
@@ -26,7 +26,7 @@ pub fn inline_fs<'a>(
 ) -> impl Fold + 'a {
   InlineFS {
     filename: Path::new(filename).to_path_buf(),
-    collect: Collect::new(
+    collect: HoistCollect::new(
       source_map,
       decls,
       Mark::fresh(Mark::root()),
@@ -41,7 +41,7 @@ pub fn inline_fs<'a>(
 
 struct InlineFS<'a> {
   filename: PathBuf,
-  collect: Collect,
+  collect: HoistCollect,
   global_mark: Mark,
   project_root: &'a str,
   deps: &'a mut Vec<DependencyDescriptor>,
