+++ packages/transformers/js/core/src/lib.rs
@@ -18,6 +18,7 @@ mod env_replacer;
 mod fs;
 mod global_replacer;
 mod hoist;
+mod hoist_collect;
 mod modules;
 mod utils;
 
@@ -29,9 +30,10 @@ use path_slash::PathExt;
 use serde::{Deserialize, Serialize};
 use swc_common::comments::SingleThreadedComments;
 use swc_common::errors::{DiagnosticBuilder, Emitter, Handler};
+use swc_common::DUMMY_SP;
 use swc_common::{chain, sync::Lrc, FileName, Globals, Mark, SourceMap};
 use swc_ecma_preset_env::{preset_env, Mode::Entry, Targets, Version, Versions};
-use swc_ecmascript::ast::Module;
+use swc_ecmascript::ast::{Invalid, Module};
 use swc_ecmascript::codegen::text_writer::JsWriter;
 use swc_ecmascript::parser::lexer::Lexer;
 use swc_ecmascript::parser::{EsConfig, PResult, Parser, StringInput, Syntax, TsConfig};
@@ -41,7 +43,7 @@ use swc_ecmascript::transforms::{
   optimization::simplify::dead_branch_remover, optimization::simplify::expr_simplifier,
   pass::Optional, proposals::decorators, react, typescript,
 };
-use swc_ecmascript::visit::FoldWith;
+use swc_ecmascript::visit::{FoldWith, VisitWith};
 
 use decl_collector::*;
 use dependency_collector::*;
@@ -49,6 +51,7 @@ use env_replacer::*;
 use fs::inline_fs;
 use global_replacer::GlobalReplacer;
 use hoist::hoist;
+use hoist_collect::HoistCollect;
 use modules::esm2cjs;
 use utils::{CodeHighlight, Diagnostic, DiagnosticSeverity, SourceLocation, SourceType};
 
@@ -414,6 +417,20 @@ pub fn transform(config: Config) -> Result<TransformResult, std::io::Error> {
                 }
               }
             } else {
+              let mut symbols_collect = HoistCollect::new(
+                source_map.clone(),
+                decls,
+                Mark::fresh(Mark::root()),
+                global_mark,
+                false,
+              );
+              module.visit_with(&Invalid { span: DUMMY_SP } as _, &mut symbols_collect);
+
+              println!(
+                "{} {:?} {:?}\n\n",
+                config.filename, symbols_collect.imports, symbols_collect.exports
+              );
+
               let (module, needs_helpers) = esm2cjs(module, versions);
               result.needs_esm_helpers = needs_helpers;
               module
